backend/requirements.txt
pyttsx3==2.90
comtypes==1.2.0
gTTS==2.5.1
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.9.0
requests==2.31.0
numpy>=1.26.0
setuptools>=70.0.0
ultralytics>=8.2.0
torch>=2.6.0
torchvision>=0.19.0
transformers>=4.40.0
Pillow>=10.2.0
opencv-python>=4.8.0
ffmpeg-python==0.2.0
vosk==0.3.45
pytesseract==0.3.10
google-generativeai>=0.8.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0